pyproject.toml
[tool.poetry]
name = "supply-chain-orchestrator"
version = "0.1.0"
description = "Agentic Supply-Chain Resilience Orchestrator"
authors = ["Supply Chain Team"]
readme = "README.md"
packages = [{include = "orchestrator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web framework
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"
pgvector = "^0.3"
alembic = "^1.13"

# Config + validation
pydantic = "^2.7"
pydantic-settings = "^2.3"

# Agents
langgraph = "^0.2"
langchain = "^0.3"
langchain-anthropic = "^0.3"
langchain-openai = "^0.2"
langchain-community = "^0.3"

# LlamaIndex
llama-index-core = "^0.11"
llama-index-embeddings-openai = "^0.2"
llama-index-vector-stores-postgres = "^0.2"

# Data / Simulation
numpy = "^1.26"
pulp = "^2.8"

# HTTP + ingestion
httpx = "^0.27"
feedparser = "^6.0"
apscheduler = "^3.10"

# Utilities
structlog = "^24.1"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-httpx = "^0.30"
ruff = "^0.4"
mypy = "^1.10"
httpx = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
